remove unnessecary fonts, refactor from feedback
