+++ styles/main.css
@@ -36,7 +36,7 @@ input::-webkit-inner-spin-button {
 	font-style: normal;
 	font-weight: 400;
 	font-display: swap;
-	src: local('Inter');
+	src: local('Inter'), url('/fonts/Inter-Regular.woff2') format('woff2');
 }
 
 @font-face {
@@ -44,7 +44,7 @@ input::-webkit-inner-spin-button {
 	font-style: normal;
 	font-weight: 600;
 	font-display: swap;
-	src: local('Inter Semi Bold');
+	src: local('Inter Semi Bold'), url('/fonts/Inter-SemiBold.woff2') format('woff2');
 }
 
 @font-face {
@@ -52,7 +52,7 @@ input::-webkit-inner-spin-button {
 	font-style: normal;
 	font-weight: 700;
 	font-display: swap;
-	src: local('Inter Bold');
+	src: local('Inter Bold'), url('/fonts/Inter-Bold.woff2') format('woff2');
 }
 
 /* GT America */
@@ -62,7 +62,7 @@ input::-webkit-inner-spin-button {
 	font-style: normal;
 	font-display: swap;
 	font-weight: 700;
-	src: local('GT America Mono');
+	src: local('GT America Mono'), url('/fonts/GT-America-Mono-Bold.woff2') format('woff2');
 }
 
 @font-face {
@@ -70,7 +70,8 @@ input::-webkit-inner-spin-button {
 	font-style: normal;
 	font-display: swap;
 	font-weight: 500;
-	src: local('GT America Condensed-Medium');
+	src: local('GT America Condensed-Medium'),
+		url('/fonts/GT-America-Condensed-Medium.woff2') format('woff2');
 }
 
 @font-face {
@@ -79,7 +80,8 @@ input::-webkit-inner-spin-button {
 	font-style: normal;
 	font-weight: 700;
 	font-stretch: condensed;
-	src: local('GT America Condensed-Bold');
+	src: local('GT America Condensed-Bold'),
+		url('/fonts/GT-America-Condensed-Bold.woff2') format('woff2');
 }
 
 @font-face {
@@ -87,5 +89,6 @@ input::-webkit-inner-spin-button {
 	font-style: normal;
 	font-display: swap;
 	font-weight: 700;
-	src: local('GT America Extended Bold');
+	src: local('GT America Extended Bold'),
+		url('/fonts/GT-America-Extended-Bold.woff2') format('woff2');
 }
